通讯录 消息UI

+++ DearIMProject/UserUtils/MYUser.h
@@ -11,7 +11,7 @@ NS_ASSUME_NONNULL_BEGIN
 
 @interface MYUser : NSObject <NSCoding, NSSecureCoding>
 
-@property (nonatomic, assign) long userId;/**< 用户id  */
+@property (nonatomic, assign) long long userId;/**< 用户id  */
 @property (nonatomic, strong) NSString *name;/**<  用户名称 */
 @property (nonatomic, strong) NSString *email;/**<  邮箱 */
 @property (nonatomic, strong) NSString *os;/**< 型号  */
